+++ packages/bridge-controller/src/utils/quote.test.ts
@@ -14,6 +14,7 @@ import {
   calcSwapRate,
   calcCost,
   formatEtaInMinutes,
+  calcSlippagePercentage,
 } from './quote';
 import type {
   GenericQuoteRequest,
@@ -521,5 +522,41 @@ describe('Quote Metadata Utils', () => {
       expect(result.valueInCurrency).toBeNull();
       expect(result.usd).toBeNull();
     });
+  });
+
+  describe('calcSlippagePercentage', () => {
+    it.each([
+      ['100', '100', '100', '100', 0, 0],
+      ['95', '95', '100', '100', 5, 5],
+      ['98.3', '98.3', '100', '100', 1.7, 1.7],
+      [null, null, '100', '100', null, null],
+      ['105', '105', '100', '100', 5, 5],
+    ])(
+      'calcSlippagePercentage: calculate slippage absolute value for received amount %p, usd %p, sent amount %p, usd %p to expected amount %p, usd %p',
+      (
+        returnValueInCurrency: string | null,
+        returnUsd: string | null,
+        sentValueInCurrency: string | null,
+        sentUsd: string | null,
+        expectedValueInCurrency: number | null,
+        expectedUsd: number | null,
+      ) => {
+        const result = calcSlippagePercentage(
+          {
+            valueInCurrency: returnValueInCurrency,
+            usd: returnUsd,
+          },
+          {
+            amount: '',
+            valueInCurrency: sentValueInCurrency,
+            usd: sentUsd,
+          },
+        );
+        expect(result.percentageInCurrency).toStrictEqual(
+          expectedValueInCurrency,
+        );
+        expect(result.percentageInUsd).toStrictEqual(expectedUsd);
+      },
+    );
   });
 });